Add files via upload
Video-Forgery-Detection
I have worked on video forgery detection and method used is based on the concept of optical flow variation factor.THE method relies on the fact that envelop of the optical flow sequence in an origi- nal video is approximately continuous. The corresponding optical flow variation sequence only fluctuates in a small range. However, the forgery processes will introduce discontinuity points in the optical flow sequence, even though there are no visible traces of tampering. Note that we only con- sider surveillance videos recorded by static cameras; hence the change of scenes, which result in obvious transitions in the optical flow sequence, will not occur in the original videos.The dataset used is SULFA dataset and also some self created forged videos. Furthermore, the focus is on ‘‘meaningful’’ forgeries (i.e. the tampering operations which can change the semantic meaning) in- cluding the deletion, insertion, and duplication of successive frames which are not changeless.The various observations from the graphs which hep in detecting forgery and also telling the type of in- ter frame forgery are: • Frame deletion Only one discontinuity point can be observed in the optical flow variation sequence. • Frame insertion Two discontinuity points occur at the beginning and ending positions of the inserted frames. Note that the variation factors at the discontinuity points are larger than the one in frame deletion forgery. • Frame duplication Two discontinuity points can be observed at the break points of the duplicated frames. My motivation for further research involve two aspects. First,include machine learning techniques in order to improve the process of video forgery detection.Second ,try to detect copy-move forgery and also work in the detection of deep fakes.